Commit 4, adding stops
This commit includes everything according to commit #4, which includes all of the stop counting for each area. I would rather have a for loop asking whether the user's inputs were taken or not but I don't know how to do that even after trying to find it on my own haha.