Commit #2 With all of the Inputs
This commit includes everything that falls under commit #2 in the directions. That involves the while loops asking the player for which direction they want to go to next and then deciding the next location after that. FOR SOME REASON, the 'stops' thing bugs the entire code and I do not know what I am doing wrong/what I need to do differently.