Commit #2 Using the Dictionary
This commit encompasses everything involved with commit #2 according to the PDF. However, I actually moved everything to a new file (maselek_adventure_dictionary.py) so as to keep the original file as is in case I need it. The new file is the retooled version of my game using the dictionary.