Commit #3
This commit includes the 2d matrix mentioned in the directions for commit #3. I also moved all of the text for each area to individual variables and also made the userName variable a global. I also moved the game loop into a function as well as the intro. Also, the counter is fixed.