Commit #03 World Class
This commit includes the World Class. It also includes a mutator and observer function for the locale class, though I do not any relevant ways to include an observer or mutator function for the world class.
I also deleted the old version of maselek_adventure.py to declutter the overall file.