Commit #04 player.py
This commit includes everything with the player class for the adventure game. I'm not sure whether the code is currently using the movement_list, areas list, and was_visited correctly though I am pretty sure it is?

Also ther was a couple of things i needed to change in world.py, such as adding commas in between each individual list and adding notes to better clarify each part of the matrix.