Commit #05 adventure but not working?
This commit includes the addition of the classes into the main code. However, it unfortunately doesn't function, and after a some bug testing I do not know how to go about fixing it. Sorry.

I also released I was missing an examine method in player and I also needed to fix the way that player accessed the movement matrix. I also had to edit a way to "end" the game.