Project 5, Commit #1
In this commit is all the code surrounding adding NPCs to the game. I added 3 NPCs, one of which when talked to gives the answer to the question the last NPC asks (which gives you score).

Also in this commit is my failed attempt at adding GUI instead of adding NPCs. haha